chipsroofing-prod/requirements/build.txt
-r common.txt
django-jenkins==0.16.2
pylint==1.3.0